third_party/WebKit/LayoutTests/svg/filters/feBlend-invalid-mode-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: Failed to set the 'baseVal' property on 'SVGAnimatedEnumeration': The enumeration value provided (65535) is larger than the largest allowed value (5).
PASS: Invalid blend modes do not trigger a crash. 
